ylamsiah

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -22,5 +22,5 @@ int	_atoi(char *s)
 			break;
 		s++;
 	}
-	return ((int )n * sign);
+	return (n * sign);
 }
